+++ R/check_data.R
@@ -105,29 +105,6 @@ check_strata <- function(strata_data) {
   }
 }
 
-#' Check Stratification Levels
-#'
-#' Check input format for stratification variable(s).
-#'
-#' @param strata_data - data.frame or character vector or factor vector
-#' @importFrom utils.nest is_character_vector
-#'
-#' @noRd
-check_strata_levels <- function(strata_data) {
-
-  if (is_character_vector(strata_data) || is.factor(strata_data)) {
-    strata_levels <- length(unique(strata_data)) > 1
-  }
-  else if (is.data.frame(strata_data)) {
-    strata_levels <- vapply(strata_data, FUN = function(x) {
-      length(unique(x)) > 1
-    }, logical(1))
-  }
-
-  if (!all(strata_levels)) {
-    stop("Not all strata variables have more than 1 level.")
-  }
-}
 
 #' Checks Numeric Ranges
 #'