+++ NEWS.md
@@ -4,6 +4,7 @@ This version of `tern` introduces a major rewriting of `tern` due to the change 
 
 * Counting patients with abnormal values post-baseline: `s_count_abnormal` and `analyze_abnormal`.
 * Summarize variables: S3 generic `s_summary`, `s_summary.numeric` method, `summarize_vars`.
+* Estimate difference in terms of responder proportions: `estimate_proportion_diff`, `s_proportion_diff`.
 
 * issues with no news: